python-worker/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
httpx>=0.27.0
requests>=2.32.3
python-dotenv>=1.0
schedule>=1.2
pydantic>=2.7.4,<3
prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=7.0.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
psycopg2-binary>=2.9.9
pytest-asyncio>=0.23.7
